requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
ca-certificates=2024.9.24=h06a4308_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
filelock=3.16.1=pypi_0
fsspec=2024.10.0=pypi_0
h5py=3.11.0=pypi_0
huggingface-hub=0.26.2=pypi_0
idna=3.10=pypi_0
joblib=1.4.2=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
metric-learn=0.7.0=pypi_0
ncurses=6.4=h6a678d5_0
numpy=1.24.4=pypi_0
nvidia-ml-py=12.535.161=pypi_0
nvitop=1.3.2=pypi_0
openssl=3.0.15=h5eee18b_0
packaging=24.1=pypi_0
pandas=2.0.3=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=py38h06a4308_0
pot=0.9.4=pypi_0
psutil=6.1.0=pypi_0
python=3.8.20=he870216_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
quadprog=0.1.12=pypi_0
readline=8.2=h5eee18b_0
requests=2.32.3=pypi_0
safetensors=0.4.5=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
setuptools=75.1.0=py38h06a4308_0
six=1.16.0=pypi_0
sqlite=3.45.3=h5eee18b_0
termcolor=2.4.0=pypi_0
threadpoolctl=3.5.0=pypi_0
timm=1.0.11=pypi_0
tk=8.6.14=h39e8969_0
torch=1.8.1+cu111=pypi_0
torchaudio=0.8.1=pypi_0
torchvision=0.9.1+cu111=pypi_0
tqdm=4.66.6=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=pypi_0
wheel=0.44.0=py38h06a4308_0
xz=5.4.6=h5eee18b_1
zlib=1.2.13=h5eee18b_1
